09.other.projects/04.tagging/requirements.txt
mutagen>=1.45.1
Pillow>=10.0.0
# Note: ffmpeg needs to be installed on the system 